DATA-PROJECTS/DATA-PROJECT-3/estado_data_generator/requirements.txt
google-cloud-firestore
google-cloud-bigquery
flask
functions-framework
python-dotenv
psycopg2-binary